more changes applied

+++ task_05/src/RMQ.cpp
@@ -23,5 +23,5 @@ int RMQ::findMin(const unsigned int left, const unsigned int right){
     if(left > right || left < 0 || right > minimum_table[0].size()) 
         throw std::out_of_range ("wrong border arguments");
     const size_t row = std::floor(std::log2(right - left + 1));
-    return static_cast<int>(std::min(minimum_table[row][left], minimum_table[row][right - std::pow(2, row) + 1]));
+    return static_cast<int>(std::min(minimum_table[row][left], minimum_table[row][static_cast<size_t>(right - std::pow(2, row) + 1)]));
 }
